Fix a bug in cover when used with no_auto_import
Cover was rewriting guard clauses as non-remote calls.
For instance, if a parse transform adds a guard clause
as erlang:is_binary(Binary), Cover was incorrectly removing
the erlang prefix which lead to errors if is_binary is
not auto imported.

This commit keeps the abstract format as it is.